Extract time funcs to a new module
